third_party/WebKit/LayoutTests/fast/dom/content-element-select-null.html

<!DOCTYPE HTML>
<title>content element test for select attribute, null and undefined arguments</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

test(function() {
    var element = document.createElement('content');
    assert_equals(element.select, '');
    element.select =  null
    assert_equals(element.select, 'null');
    element.select =  undefined
    assert_equals(element.select, 'undefined');
});
</script>
